Enable BCJ fiter support in xz-embedded.
BCJ (branch/call/jump) filters in XZ will convert relative jumps into
absolute addresses to increase the redundancy. Since binary code has
plenty of relative jumps, this often improves compression ratio of .xz.
Preliminary test shows an improvement of 2% in the compression ratio
when compressing ELF files, which is already around 25% compression
ration when using xz.

This patch enables the BCJ for x86, arm and thumb code filters (the
actual data being compressed, not the supported architectures where
the filter can run). This increases the size of the .a in ~12 KiB.

Bug: 27817327
TEST=Used xz-embedded to decompress a file compressed with `xz --x86 --lzma2`

Change-Id: Iadb9dd8174ba404a6dfa8e9bddecd878af4b2fcf