Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> +cferris for the external/libunwind build failure:
> 
>   external/libunwind/src/elfxx.c:472:15: error: incompatible pointer types assigning to 'void *(*)(ISzAllocPtr, size_t)' (aka 'void *(*)(const struct ISzAlloc *, unsigned long)') from 'void *(void *, size_t)' (aka 'void *(void *, unsigned long)') [-Werror,-Wincompatible-pointer-types]
> 
> can we remove external/libunwind yet?
> 
> if not, i think we can just `#undef HAVE_LZMA` in https://cs.corp.google.com/android/external/libunwind/include/config.h?type=cs&q=HAVE_LZMA+file:%5Eexternal/libunwind/+package:%5Eandroid$&g=0&l=134
> 
> +hhb as "mr external". (i've added lzma to the spreadsheet.)

I thought there were a few places that still actually depended on it. I filed a b/79209464 to remove this.

Last time I tried to remove it, it didn't work, but I didn't notice that there are a few places that appear to be including it, but aren't using it. I'm going to remove those first.

Patch-set: 1
Reviewer: Christopher Ferris <1019050@85c56323-6fa9-3386-8a01-6480fb634889>