Update patch set 3
Patch Set 3: Code-Review+2

> Patch Set 3:
> 
> > Patch Set 2:
> > 
> > This should use Merged-In, not DO NOT MERGE (which would cause it to be reverted in the next push)
> 
> Doesn't this cause the same effect?

No, DO NOT MERGE would cause this change to be skipped in stage-aosp-master and pi-dev-plus-aosp, so it would only end up in aosp-master and master. The next time we push stage-aosp-master to aosp-master, the change would be essentially reverted there as well.

In general, always use Merged-In, it's significantly safer and more predictable.

Patch-set: 3
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2